rename component part 2
