+++ src/app/nip76/nip76-main/nip76-main.component.ts
